+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Features
 
-- Adds option `Compress` for in-memory gzip compression (#18)
+- Adds option `Compress` for in-memory gzip compression (#18, #19)
 
 ## [[v0.1.3]](https://github.com/mlange-42/ark-serde/compare/v0.1.2...v0.1.3)